requirements.txt
Sphinx>=1.4.1
sphinx-rtd-theme>=0.1.9
sphinx-autodoc-typehints>=1.2.5,!=1.5.0
pytest>=3.4.2
pytest-mock>=1.7.1
future
ruamel.yaml>=0.12.10
schema>=0.6.2
behave>=1.2.6
typing>=3.5.1
six
backports.tempfile